+++ R/aim.R
@@ -14,79 +14,84 @@
 #'    + \sum_{k} \gamma_{k} f_{k}(x_{ik}) + e_{i}}
 #'  The formula interface considers \code{\link{g}} to identify index terms,
 #'   \code{\link{s}} for smooth functions and can also include 
-#'    linear terms. All smooth terms can be
-#'    shape constrained.
+#'    linear terms. All smooth terms can be shape constrained.
 #'
 #' The CGAIM allows for linear constraints on the alpha coefficients. 
-#'  Common constraints can be given for each index through the 
-#'  \code{\link{g}} terms in the formula. Custom constraints, including multi-index constraints can be passed as a matrix through \code{alpha.control$Cmat}. See 
-#'  \code{\link{alpha.setup}}.
+#'  Common constraints can be given for each index through the \code{acons} 
+#'  parameter in \code{\link{g}} terms in the formula. Custom constraints 
+#'  can either be given for each idnex individually through the 
+#'  \code{\link{g}} interface, or for all parameters through
+#'  \code{alpha.control$Cmat}. See \code{\link{alpha.control}}.
 #'
 #' The CGAIM is fitted through an iterative algorithm that alternates between
 #'  estimating the ridge functions \eqn{g_{j}} (and other non-index terms) and 
 #'  updating the coefficients \eqn{\alpha_{j}}. The smoothing of ridge functions
-#'  currently supports three methods: \code{\link[scam]{scam}} (the default), \code{\link[cgam]{cgam}} and \code{\link[scar]{scar}}. The list \code{smooth.control} allows controlling the functions.
+#'  currently supports three methods: \code{\link[scam]{scam}} (the default), 
+#'  \code{\link[cgam]{cgam}} and \code{\link[scar]{scar}}. 
+#'  The list \code{smooth.control} allows controlling the functions.
 #'
 #' Updating the coefficient is made through quadratic programming. Currently,
 #'  it is carried out by either the function 
 #'  \code{\link[osqp:solve_osqp]{solve_osqp}} or 
-#'  \code{\link[quadprog:solve.QP]{solve.QP}}. Controlling the updates is
-#'  made through the parameter \code{alpha.control}. See \code{\link{alpha.setup}}
-#'  for the detail.
+#'  \code{\link[quadprog:solve.QP]{solve.QP}}. See \code{\link{alpha.control}}
+#'  for details.
 #'
 #' @param formula A CGAIM formula with index terms \code{\link{g}}, 
-#'    smooth terms \code{\link{s}} and linear terms. 
+#'  smooth terms \code{\link{s}} and linear terms. 
 #' @param data A data.frame containing the variables of the model.    
-#' @param weights A vector of optional weights for observations. If \code{NULL}, unit weights are considered.
+#' @param weights A vector of optional weights for observations. 
 #' @param na.action A function indicating what treatment applying to NAs. If
-#'    missing the default is set by the \code{na.action} setting of \code{options}. See
-#'    \code{\link[stats]{na.fail}}.
-#' @param smooth_method The shape constrained smoothing method to consider. The default is \code{\link[scam]{scam}}, but other options are \code{\link[cgam]{cgam}} and \code{\link[scar]{scar}}. See details.
-#' @param smooth.control A list containing parameters for the shape-contrained smoothing function given passed to \code{smooth_method}. See help of the functions for the list of parameters.
-#' @param alpha.control A list containing the controlling parameters for the
-#'    alpha optimization steps of the algorithm. See \code{\link{alpha.setup}}.
-#' @param algo.control A list containing controlling parameters for the
-#'    whole algorithm. See \code{\link{algo.setup}}.
-#' @param keep.trace Logical. If TRUE, the result contains a 'trace' element
-#'    giving the intermediate values of alpha coefficients and ridge functions
-#'    at each step of the algorithm. Useful for behaviour tracking.
+#'  missing the default is set by the \code{na.action} setting of \code{options}. 
+#'  See \code{\link[stats]{na.fail}}.
+#' @param smooth_method The shape constrained smoothing method to consider. 
+#'  The default is \code{\link[scam]{scam}}, but other options are 
+#'  \code{\link[cgam]{cgam}} and \code{\link[scar]{scar}}. See details.
+#' @param smooth_control A list of control parameters for the shape-constrained 
+#'  smoothing function given in \code{smooth_method}. 
+#'  See help of the functions for the list of available parameters.
+#' @param alpha_control A list of parameters controlling the
+#'    alpha updating steps of the algorithm. See \code{\link{alpha.control}}.
+#' @param algo_control A list containing controlling parameters for the
+#'    outer algorithm. See \code{\link{algo.control}}.
 #'
 #' @return A \code{cgaim} object, i.e. a list with components:
 #'  \item{alpha}{A named list of index coefficients.}
 #'  \item{gfit}{A matrix containing the ridge and smooth functions 
-#'    evaluated at the observations. Note that column ordering puts indices first and covariates after.}
+#'    evaluated at the observations. 
+#'    Note that column ordering puts indices first and covariates after.}
 #'  \item{indexfit}{A matrix containing the indices evaluated at the 
 #'    observations.}
 #'  \item{beta}{A vector containing the intercept and the scale coefficient
 #'    of each ridge and smooth function. Includes the \eqn{\gamma_{k}} of
-#'    the CGAIM model above. Note that ordering puts indices first and covariates after.}.
+#'    the CGAIM model above. Note that ordering puts indices first and 
+#'    covariates after.}.
 #'  \item{index}{A vector identifying to which index the columns of the
 #'    element \code{x} belong.}
-#'  \item{fitted}{A vector of fitted y values.}
+#'  \item{fitted}{A vector of fitted responses.}
 #'  \item{residuals}{A vector of residuals.}
 #'  \item{rss}{The residual sum of squares of the fit.}
 #'  \item{flag}{A flag indicating how the algorithm stopped. 1 for proper 
-#'    convergence, 2 when the algorithm stopped for lack of step in the
-#'    descent direction and 3 when the maximum number of iterations has
+#'    convergence, 2 when the algorithm stopped for failing to decrease rss
+#'    and 3 when the maximum number of iterations has
 #'    been reached.}
-#'  \item{trace}{If \code{keep.trace = TRUE}, a tracking list containing,
-#'    for each iteration, values of alpha coefficients, 
-#'    ridge and smooth function, convergence criteria, step length,
-#'    as well as the number of iterations.}
+#'  \item{niter}{Number of iterations performed.}
+#'  \item{edf}{Effective degrees of freedom of the estimator.}
+#'  \item{gcv}{Generalized cross validation score.}
 #'  \item{covariates}{A data.frame containing the values of the variables not 
 #'    entering any index.}
 #'  \item{x}{A matrix containing the variables entering the indices. 
-#'    The variables are mapped to each index throught the element \code{index}.}
+#'    The variables are mapped to each index through the element \code{index}.}
 #'  \item{y}{The response vector.}
 #'  \item{weights}{The weights used for estimation.}
-#'  \item{smooth.control}{The controlling parameters passed to the function
-#'    for the smoothing steps.}
-#'  \item{alpha.control}{The controlling parameters passed to the function
-#'    for the alpha update steps.}
-#'  \item{algo.control}{The controlling parameters passed to the function
-#'    for the minimization algorithm.}
+#'  \item{smooth_control}{List of parameters controlling the smoothing step.}
+#'  \item{alpha_control}{List of parameters controlling alpha updating step.}
+#'  \item{algo_control}{List of parameters controlling the algorithm.}
 #'  \item{terms}{A \code{\link[stats]{terms.object}} representing the model.
 #'    Useful for prediction.}
+#'  \item{trace}{If \code{algo_control$trace = TRUE}, a tracking list 
+#'    containing, for each iteration, values of alpha coefficients, 
+#'    ridge and smooth function, convergence criteria, step length,
+#'    as well as the number of iterations.}
 #'
 #' @seealso \code{\link{confint.cgaim}} for confidence interval,
 #'    \code{\link{predict.cgaim}} to predict new data,
@@ -94,137 +99,78 @@
 #'
 #' @export
 cgaim <- function(formula, data, weights, na.action, 
-  smooth_method = c("scam", "cgam", "scar"), smooth.control = list(), 
-  alpha.control = list(), algo.control = list(), keep.trace = F) 
+  smooth_method = "scam", smooth_control = list(), 
+  alpha_control = list(), algo_control = list()) 
 {
+  # Terms
   mt <- stats::terms(formula, specials = c("g", "s"), data = data)
-  allvars <- all.vars(formula)
   gind <- attr(mt, "specials")$g
   p <- length(gind)
   ptot <- length(attr(mt, "term.labels"))
-  if (missing(na.action)) na.action <- getOption("na.action")
-  # Extract info on indices
-  data <- stats::model.frame(stats::reformulate(allvars), data = data, 
-    na.action = na.action)  
-  index_interp <- stats::model.frame(mt[gind - 1], data = data)
-  y <- stats::model.response(index_interp)
-  attr(y, "varname") <- allvars[1]
-  index_interp <- index_interp[-1]
-  # Labels for smoothing step
-  index_labels <- sapply(index_interp, attr, "label")
-  ulabs <- unique(index_labels)
-  plabs <- length(ulabs)
-  if (plabs < p){
-    for (ilab in ulabs){
-      wlabs <- which(index_labels == ilab)
-      if (length(wlabs) > 1){
-        index_labels[wlabs] <- sprintf("%s%i", ilab, 1:length(wlabs))
-      }
-    }
-  }
-  # Organizing variables in indices as a matrix with index positions   
-  pvec <- sapply(index_interp, attr, "nterms")
-  index <- rep(1:p, pvec)
-  names(index) <- rep(index_labels, pvec)
-  n <- nrow(index_interp)
-  Xind <- matrix(0, n, sum(pvec))
-  for (j in 1:p){
-    Xind[,index == j] <- index_interp[[j]]
-    colnames(Xind)[index == j] <- colnames(index_interp[[j]])
-  }
-  # Initialize alpha controls
-  alpha.control <- do.call(alpha.setup, alpha.control)
-  nai <- length(alpha.control$alpha.start)
-  if (nai > 0 && nai != sum(pvec)){
-    warning("alpha.init length is inconsistent with index matrix and is recycled")
-    alpha.control$alpha.start <- rep_len(alpha.control$alpha.start, sum(pvec))
-  }
-  # Organizing the constraint matrix
-  gconsts <- as.matrix(Matrix::bdiag(lapply(index_interp, attr, "Cmat")))
-  constr_mat <- alpha.control$Cmat
-  if (!is.null(constr_mat)){
-    if (ncol(constr_mat) != ncol(Xind)){
-      stop("Number of columns in alpha.control$Cmat does not match
-        the indices")
-    }
-  }
-  if (length(gconsts) > 0 || length(constr_mat) > 0){
-    Cmat <- rbind(constr_mat, gconsts)
-    Cmat <- Cmat[!duplicated(Cmat),, drop = F]
-    alpha.control$Cmat <- Cmat
-  }
+  # Fill missing arguments
+  # n <- NROW(data[[1]])
+  # if (is.null(weights)){
+  #   weights <- rep(1, n)
+  # } else {
+  #   weights <- rep_len(weights, n)
+  # }
+  # if (is.null(na.action)) na.action <- getOption("na.action")
+  # Extract data
+  mf <- match.call(expand.dots = FALSE)
+  m <- match(c("data", "weights", "na.action"), 
+    names(mf), 0L)
+  mf <- mf[c(1L, m)]
+  mf$drop.unused.levels <- TRUE
+  mf$formula <- mt
+  mf[[1L]] <- quote(stats::model.frame)
+  mf <- eval(mf, parent.frame())
+  # mf <- stats::model.frame(mt, data = data, 
+  #   na.action = na.action, weights = weights)
+  # Initialize index estimation components
+  mod_components <- index.setup(mf, alpha_control)
   # Prepare parameters for smoothing
-  smooth_method <- match.arg(smooth_method)
-  setup_fun <- sprintf("%s.setup", smooth_method)
-  smooth.control <- do.call(setup_fun, list(mt, data, smooth.control))
-  algo.control$smooth_method <- smooth_method
-  # Calling the fitting function
-  pars <- algo.control[names(algo.control) %in% 
-    methods::formalArgs(gaim_gn)]
-  pars$x <- Xind
-  pars$y <- y
-  if (missing(weights)){
-    pars$w <- rep(1, n)
-  } else {
-    pars$w <- rep_len(weights, n)
-  }
-  pars$index <- index
-  pars$smooth.control <- smooth.control
-  pars$alpha.control <- alpha.control
+  smooth_components <- smooth.setup(mf, smooth_method, smooth_control)
+  # Prepare parameters for fitting algorithm
+  algo_control <- do.call(algo.control, algo_control)
   # Fitting the model
-  result <- do.call(gaim_gn, pars)
-  result$alpha <- split(result$alpha, index)
-  for (j in 1:p){
-    names(result$alpha[[j]]) <- colnames(index_interp[[j]])
-  }
-  names(result$alpha) <- index_labels
+  algo_pars <- c(mod_components, algo_control, 
+    list(smooth_control = smooth_components))
+  result <- do.call(gaim_gn, algo_pars)
+  # Organize output
+  result$alpha <- split(result$alpha, mod_components$index)
+  names(result$alpha) <- unique(names(mod_components$index))
   attributes(result$gfit) <- attributes(result$gfit)[c("dim", "dimnames")]
-  if (!is.null(result$bvcov)){
-    colnames(result$bvcov) <- rownames(result$bvcov) <- names(result$beta)
-  }  
-  result$covariates <- smooth.control$Xcov
-  result$x <- Xind
-  result$y <- y
-  result$weights <- pars$w
-  result$smooth.control <- smooth.control
-  result$alpha.control <- alpha.control
-  result$algo.control <- algo.control[names(algo.control) %in% 
-    methods::formalArgs(gaim_gn)]
+  result$covariates <- smooth_components$Xcov
+  result$x <- mod_components$x
+  result$y <- mod_components$y
+  result$weights <- mod_components$w
+  result$smooth_control <- smooth_components
+  result$alpha_control <- mod_components$alpha_control
+  result$algo_control <- algo_control
   result$terms <- mt
   class(result) <- "cgaim"
   return(result)
 }
 
 gaim_gn <- function(x, y, w, index, 
-  smooth.control = list(), alpha.control = list(),
-  keep.trace = FALSE, max.iter = 50, tol = 1e-3, min.step.len = 0.1, 
-  halving = T, convergence_criterion = c("change", "offset"), 
-  smooth_method = "scam")
+  smooth_control = list(), alpha_control = list(),
+  trace = FALSE, max.iter = 50, tol = 1e-3, min.step.len = 0.1, 
+  halving = T, convergence_criterion = c("change", "offset"))
 {
-  convergence_criterion <- match.arg(convergence_criterion)
-  if (convergence_criterion == "change") tol <- rep_len(tol, 2)
   # Useful objects
   n <- length(y)
   p <- max(index)
   d <- length(index)
   ind_pos <- split(1:d, index)
-  # Alpha initialization
-  if (is.null(alpha.control$alpha.start)){    
-    init_pars <- alpha.control
-    init_pars <- within(init_pars,{
-      y = y; x = x; w = w; index = index
-    })
-    alpha <- do.call(alpha_init, init_pars)
-  } else {
-    alpha <- alpha.control$alpha.start
-  }
   # Indice computation
+  alpha <- alpha_control$alpha
   zs <- sapply(ind_pos, function(i, x, a) x[,i] %*% a[i], 
     x = x, a = alpha)
   colnames(zs) <- unique(names(index))
   # Initial smoothing
-  smooth_fun <- sprintf("smooth_%s", smooth_method) 
-  smo_par <- c(smooth.control, 
+  smooth_fun <- sprintf("smooth_%s", smooth_control$method)
+  smooth_control$method <- NULL
+  smo_par <- c(smooth_control, 
     list(y = y, x = zs, weights = w))
   gz <- do.call(smooth_fun, smo_par)   
   yhat <- gz$intercept + rowSums(gz$gz)
@@ -234,7 +180,7 @@ gaim_gn <- function(x, y, w, index,
   c1 <- 1      
   l2 <- L2(y, yhat, w)  
   # If requested: tracing the algorithm evolution
-  if (keep.trace){
+  if (trace){
     trace.list <- list(
       criterion = matrix(NA, max.iter, length(tol)),
       alpha = matrix(NA, nrow = max.iter, ncol = d),
@@ -248,8 +194,7 @@ gaim_gn <- function(x, y, w, index,
   }
   # Gauss-Newton search
   alpha_pars <- c(
-    alpha.control[names(alpha.control) %in% 
-      methods::formalArgs(alpha_update)],
+    alpha_control[names(alpha_control) %in% methods::formalArgs(alpha_update)],
     list(x = x, w = w, index = index, delta = TRUE))
   stopflag <- 0
   while(stopflag == 0){
@@ -257,14 +202,15 @@ gaim_gn <- function(x, y, w, index,
     alpha_pars$y <- r
     alpha_pars$dgz <- gz$dgz
     alpha_pars$alpha <- alpha
-    delta <- do.call(alpha_update, alpha_pars)
+    alphaup <- do.call(alpha_update, alpha_pars)
+    delta <- alphaup$alpha
     # Halving in case of bad steps
     cuts <- 1
     repeat {   
       delta <- delta * cuts
       alpha.new <- alpha + delta
-      alpha.new <- unlist(
-        tapply(alpha.new, index, normalize, type = alpha.control$norm.type))
+      alpha.new <- unlist(tapply(alpha.new, index, normalize, 
+        type = alpha_control$norm.type))
       zs <- sapply(ind_pos, function(i, x, a) x[,i] %*% a[i], 
         x = x, a = alpha.new)
       colnames(zs) <- unique(names(index))   
@@ -290,7 +236,7 @@ gaim_gn <- function(x, y, w, index,
     alpha <- alpha.new
     l2 <- l2.new
     c1 <- c1 + 1       
-    if (keep.trace){ # Tracing the algorithm evolution
+    if (trace){ # Tracing the algorithm evolution
       trace.list$criterion[c1,] <- eps
       trace.list$step.len[c1] <- cuts 
       trace.list$alpha[c1,] <- alpha 
@@ -301,21 +247,25 @@ gaim_gn <- function(x, y, w, index,
     } else { 
       if (c1 >= max.iter){
         stopflag <- 3
-        warning(sprintf("Fitting did not converge after %i iterations. Consider revising the constraints", max.iter))
+        warning(paste0("Fitting did not converge after ", max.iter, 
+          " iterations. Consider revising the constraints"))
       } 
     }    
   }
+  names(alpha) <- colnames(x)
   final.gz <- scale(gz$gz)
   final.gz[,apply(gz$gz, 2, stats::sd) == 0] <- 
     gz$gz[,apply(gz$gz, 2, stats::sd) == 0]
   betas <- attr(final.gz, "scaled:scale")
   r <- y - yhat
-  sig2 <- sum(r^2) / (n - d)
+  edf <- gz$edf + d - sum(alphaup$active)
+  gcv <- l2 / (1 - (edf / n))^2
+  sig2 <- sum(r^2) / (n - edf)
   output <- list(alpha = alpha, gfit = final.gz, indexfit = zs, 
     beta = c(gz$intercept + sum(attr(final.gz, "scaled:center")), betas),
     index = index, fitted = yhat, residuals = r, rss = l2, flag = stopflag, 
-    niter = c1)
-  if (keep.trace){
+    niter = c1, edf = edf, gcv = gcv)
+  if (trace){
     if (convergence_criterion == "change"){
       colnames(trace.list$criterion) <- c("rss", "alpha")
     } else {